Implemented parameters per province base code.

+++ src/data_structures/params.rs
@@ -16,7 +16,7 @@ fn hand_washing(parameters: &SimulationParameters, previous: &Vec<f32>, previous
 /// Social distancing reduces transmission by having more distance between people and limits visits etc.
 fn social_distancing(parameters: &SimulationParameters, previous: &Vec<f32>, previous_data: &[Vec<f32>], time: f32, h: f32) -> f32 {
     let delayed_hospitalizations = previous_data[previous_data.len() - ((parameters.incubation_period_in_days as f32 / h) as usize)][6];
-    return if delayed_hospitalizations >= 0.1 * parameters.max_hospital_capacity {
+    return if delayed_hospitalizations >= 0.1 * parameters.max_hospital_capacity as f32 {
         return 0.1
     } else { 0.0 }
 }
@@ -25,7 +25,7 @@ fn social_distancing(parameters: &SimulationParameters, previous: &Vec<f32>, pre
 fn soft_lock_down(parameters: &SimulationParameters, previous: &Vec<f32>, previous_data: &[Vec<f32>], time: f32, h: f32) -> f32 {
     let delayed_hospitalizations = previous_data[previous_data.len() - ((parameters.incubation_period_in_days as f32 / h) as usize)][6];
 
-    return if delayed_hospitalizations >= 0.3 * parameters.max_hospital_capacity {
+    return if delayed_hospitalizations >= 0.3 * parameters.max_hospital_capacity as f32{
         0.4
     } else { 0.0 }
 }
@@ -35,13 +35,10 @@ fn hard_lock_down(parameters: &SimulationParameters, previous: &Vec<f32>, previo
 
     let delayed_hospitalizations = previous_data[previous_data.len() - ((parameters.incubation_period_in_days as f32 / h) as usize)][6];
 
-    return if delayed_hospitalizations >= 0.6 * parameters.max_hospital_capacity {
+    return if delayed_hospitalizations >= 0.6 * parameters.max_hospital_capacity as f32 {
         0.2
     } else { 0.0 }
 }
-
-
-
 
 pub struct SimulationParameters {
     pub time_span_in_days: usize,
